feat: added authmiddleware and added it to listcontroller

+++ controllers/listController.ts
@@ -1,6 +1,7 @@
 import {Request, Response } from "express";
 import List from "../models/listModel";
 import User from "../models/userModel";
+import { AuthRequest } from "../interfaces/userInterface";
 
 export const getLists = async (req: Request, res: Response) => {
     try {
@@ -12,7 +13,6 @@ export const getLists = async (req: Request, res: Response) => {
             res.status(500).json({error: error.message});
             return;
         }
-       
     }
 }
 
@@ -34,11 +34,14 @@ export const getListByUserId = async (req:Request, res: Response) => {
     }
 }
 
-export const createList = async (req:Request, res: Response) => {
+export const createList = async (req: AuthRequest, res: Response) => {
     try {
+        const user = await User.findOne({ _id: req.userId})
+
+        const userId = req.userId;
         const { title, description } = req.body;
-        const userId = req.params.id;
-        const newList = new List ({ userId, title, description });
+        const username = user?.name;
+        const newList = new List ({ userId, username, title, description });
         await newList.save();
         res.status(201).json({newList});
 
@@ -50,10 +53,21 @@ export const createList = async (req:Request, res: Response) => {
     
     }  
 }
-export const updateList = async (req: Request, res: Response) => {
+export const updateList = async (req: AuthRequest, res: Response) => {
     try {
-        //add auth
+        const user = await List.findById(req.params.id);
+        
+        if(!user){
+            res.status(404).json({message: "List not found"});
+            return;
+        }
+
+        if (user.userId !== req.userId) {
+            res.status(403).json({message: "No access"});
+            return;
+        }
         const list = await List.findByIdAndUpdate(req.params.id, req.body, {new: true});
+
         if(!list) {
             res.status(404).json({message: "List not found"});
             return;
@@ -66,11 +80,23 @@ export const updateList = async (req: Request, res: Response) => {
         }
     }  
 }
-export const deleteList = async (req: Request, res: Response) => {
+export const deleteList = async (req: AuthRequest, res: Response) => {
 try {
+    const user = await List.findById(req.params.id);
+
+    if(!user){
+        res.status(404).json({message: "List not found"});
+        return;
+    }
+
+    if (user.userId !== req.userId) {
+        res.status(403).json({message: "No access"});
+        return;
+    }
     const list = await List.findByIdAndDelete(req.params.id);
     if (!list) {
         res.status(404).json({message: "List not found"});
+        return;
     }
     res.json({message: "List deleted."});
 } catch (error: unknown) {
